Handle pagination
My initial approach for this was gonna be "look for a 'next page' link
on the page, scrape that URL, and keep going until we get to a page that
doesn't have a 'next page link'". Simple.

However! Very annoyingly, this Greenhouse page doesn't use regular `<a>`
links for its pagination. Instead it uses a React component that renders
`<button>`s, which don't expose the page URLs in any way that Nokogiri
can access. So those pagination buttons are completely useless to us!
Gah!!!

So instead, this code manually adds the `?page=` query parameter to the
page URL, and increments it until we reach a page that doesn't have any
job listings. It achieves the same thing, just not in the way I'd prefer
(which is to behave like a real user browsing through the site).