Capture department name for each job
Not the prettiest way of grabbing this information, but it's what we
gotta do given the document tree we have to work with.

Technically I think this is more of a "sub-department" than a
"department" - the page is split into top-level departments (e.g.
"Engineering", "Marketing"), and sub-departments under that (e.g.
"Core Development", "Internal Infrastructure"), and it's that
sub-department name that we're grabbing here. However I'm eventually
going to filter this scraper to only look at jobs under the
"Engineering" department (by adding a `?department=` query parameter to
the URLs we're fetching), so that top-level department information is
irrelevant to us anyway since it'll be the same for every job post we'll
find.