Save new job listings to SQLite database
Very basic database class here that only does the two things we need it
to do:

1. Loads all existing jobs from the database and tells us their URLs, so
   that we can figure out which of the jobs we've just fetched are new.
2. Persists a set of jobs to the database.

This is based on the database code from an existing scraper I created a
few years ago (https://morph.io/blimpage/sash_windows). That used the
scraperwiki gem, which it turns out is no longer maintained. So instead
I'm directly using the sqlite3 gem that it was using under the hood.
That means I've gotta handle a few things that scraperwiki would've
automatically handled for me:

- Specifying the database filename (must be `data.sqlite` in order to
  run on morph.io)
- Creating the "data" table if it doesn't already exist (must be named
  "data" in order to run on morph.io)
- Managing the database schema (scraperwiki just kinda magically did
  this on the fly when you handed it a hash of data to persist)

So I've gone for the most basic possible approach to doing all of those
things!

Also, the addition of a bunch of platform-specific information to
Gemfile.lock here are the result of me following the advice of this
warning I got when installing the sqlite3 gem:

```
The following platform specific gems are getting installed, yet the lockfile includes only their generic ruby version:
 * sqlite3-2.6.0-arm64-darwin
Please run `bundle lock --normalize-platforms` and commit the resulting lockfile.
Alternatively, you may run `bundle lock --add-platform <list-of-platforms-that-you-want-to-support>`
```