Exclude "New" tag text from job titles
When new job ads appear, a "New" tag is displayed next to the job title.
The markup for that looks like this:

```
<p class="body body--medium">
  Fullstack Engineer
  <span class="tag-container" style="margin-left: 8px;">
    <span class="ellipse">
      <span class="tag-text">
        New
      </span>
    </span>
  </span>
</p>
```

Because we extract the job title by grabbing the inner text of that
outermost `<p>` element, we end up getting "New" appended to the end of
the job title. That's not what we want.

It doesn't seem like Nokogiri has an easy way for us to grab the inner
text of the `<p>` while excluding the text of its descendants, so
instead we now delete any `.tag-container` descendants before grabbing
the inner text. A little janky, but it does the job fairly succinctly.