Send notification email about new job listings
This code to send an email using the Sendgrid API is also cribbed from a
previous scraper I wrote, though I've cleaned it up a bit here.

Had to install the base64 gem too because it's apparently no longer part
of the Ruby standard library as of Ruby 3.4, which is what I'm running.
Presumably the sendgrid-ruby gem will get updated to list it as a
dependency at some point, and I'll be able to remove it from the Gemfile
here.