Downgrade to Ruby 2.6.2
I got this working with the latest version of Ruby (3.4.2), which
includes the latest version of Bundler (2.6.2). All was good. However
when I pushed it up to morph.io and tried to run it there, the app
failed to build. It seems like morph's Heroku setup is happy to use Ruby
3.4.x, but wants to use Bundler 1.15.2, which is [checks notes] 8 years
old.

I do have an older scraper which is running just fine, so I checked what
versions of Ruby and Bundler it's using, which is Ruby 2.6.2 and Bundler
1.17.2. So here's me downgrading this project to use those old,
unsupported versions 😬. That also required downgrading the nokigiri and
sqlite3 gems, since their latest versions have dropped support for Ruby
2.6.

I've also removed the base64 gem from the Gemfile since it's included as
a default gem in Ruby 2.6.2.

Everything works locally, so I'll push this up and see if it runs on
morph.io.

Obviously this is in no way ideal, but I'm trying to be pragmatic - this
is a project that probably only needs to exist for a few weeks, and if I
can get it running on morph.io with some old dependencies, then that's a
much quicker solution than setting it up on another hosting provider
where I'd need to configure the infrastructure, storage, and cron
schedule myself.